Print out all parts of a test which failed or which weren't checked.
Consider this test:

  VM:
    status: success

and assume it failed. Previously all we would print is the part of the test that
failed i.e. the status check: all the user would see is "Error". This is
correct, but not entirely helpful: did stderr/stdout contain useful clues?

This commit changes things so that any parts of a test which failed, or weren't
referenced, will be printed out. In the example above, that means that stderr
and stdout will also be printed out. But it also mean that if you have a test
like:

  VM:
    status: success
    stdout:

and the test fails, but with nothing printed out to stdout (i.e. in this case,
that means that the stdout test passed), then only the status and stderr will be
printed out. In other words, we try and minimise the information printed to the
user, while still giving them everything they might need to debug the problem.