Exit the process on a fatal error.
The problem with running things multi-threaded is that a panic in one thread
doesn't kill the others, so the user can easily end up not noticing that
something very bad has happened. This commit turns panics into process-exiting
fatal errors: when something bad happens, it's now very hard not to notice it!