When running 1 test thread, print out the test name before running it.
In a multi-threaded context, we print out the name after we execute the test
file (otherwise the test file and "OK"/"Failed" would get separated).

This commit changes things so that, in a single-threaded context
(--test-threads=1), we print out the name before we execute the test file. This
can help make clear if a particular test is performing oddly (slowly etc.).