Warn users when a test has exceeded a timeout.
This commit prints a warning to stderr every multiple of TIMEOUT seconds that a
given test has not completed. This is subtly different to cargo, which only
prints a warning after TIMEOUT has been exceeded once (if something's very slow,
for debugging reasons it can sometimes be nice to know *how* slow).

This is particularly useful when running parallel tests when it is often not
obvious why the test runner has not completed. This is similar to how `cargo
test` works.