Add ability to pass per test command line args
Sometimes it's desirable to run certain tests (or subcommands) with
custom command line arguments. This commit adds the ability to do this
without the need for defining a separate test harness.

We introduce a new keyword, `args`, which can be optionally specified
for each test command. For example, suppose an individual test relied on
handling some user input, one could pass this input as follows:

Run-time:
    args: 1234
    status: success

This passes the command line argument 1234 to the command named
"Run-time". Note that command line arguments are always passed as
strings.

It is important to remember that custom arguments should not break
invariants which are relied upon by the test runner. For example,
suppose a test runner has two commands: compile, and then execute a C
program. If an individual test provides a custom argument to specify a
different output for the compiled binary:

Compile:
    args: -o /tmp/
    ...

Then the test harness will fail, as it will not know to look in the
optionally specified /tmp/ directory.