Require previous commands to succeed by default
Sometimes when using lang_tester, running a command doesn't work as
below:

```
Fatal exception:
    Couldn't run command "./mybin".
```

It's frequently caused by a previous command (usually compilation)
failing, and the panic materialises in a later command as its necessary
preconditions weren't satisfied. [e.g. no binary exists!]

In a typical "`compile` and `run`" test, where the result of `run` is
tested in some way. There is an assumption by the test author that the
compilation succeeded and produced a binary. When it doesn't, the runner
panics as no executable for "run" exists. I think this is poor UX for
two reasons:

  1) Panicking prevents further tests from executing

  2) It's likely that it was a mistake on the test author's part.
  Panicking hides the bug and, if it was the effect of an erroneous
  previous command, it's difficult for the author to examine the stdout
  / stderr of those to see why.

This commit changes the default behaviour of the test runner to address
this. It assumes that if a test contains a list of commands from 1..n,
and commands 1..(n-1) have no explicit test cases, they *must* succeed
before the next command is run. If this is undesired, individual
commands should be given explicit test-cases to verify their desired
output. In addition, entire tests will still be ignored if no test cases
are provided.