Add status for specifying that a signal was caught
If the user expects a command, `x`, to terminate due to a signal being
thrown, they can now specify this explicitly, rather than using the
generic "error" status.

A test will still pass if it expects an "error" status from `x` which
terminated with a signal, as "error" is a superset of "signal".