Add support for `--ignored`.
This follows `cargo test`'s semantics in that it *only* runs tests marked
"ignore".

As part of this, I finally had to admit that some of the code is too deeply
nested, and part of too long a function, for me to understand. This thus breaks
one large function into two. It's still not good enough, but it's better than
before.