Print out the name of nested tests in a more pleasing way.
Although I'd never tried it, lang_tester supported nested tests (i.e. tests in
nested directories). However, if you had something like:

  lang_tests/
    a.rs
    b/
      c.rs

then the output would be:

  lang_tests::a.rs
  lang_tests::c.rs

That seems OK until you do this:

  lang_tests/
    a.rs
    b/
      a.rs

and then the output would be:

  lang_tests::a.rs
  lang_tests::a.rs

Which test is which?

This commit makes it so that nested tests print out their directory name(s):

  lang_tests::a.rs
  lang_tests::b::a.rs

This does rely on path canonicalisation giving back filenames whose prefixes
match up, which is never entirely guaranteed to succeed, so there is a fallback
mechanism if that doesn't work.