Add the ability to override fm's defaults when using lang_tester.
This allows lang_tester users to e.g. use name matching in their tests.

I'm not sure if this is the right API or not: it's flexible but that brings some
complexity as well. We could, for example, pass a boolean `is_stderr`, but that
would restrict us forever more to just stderr/stdout tests. Maybe that's all we
need, but I'm not sure.