Make test_extract take `&Path`s rather than `&str`s.
lang_tester tries to make it as easy as possible to set-up, so it has a fairly
simple API. One way I did that was by passing the contents of a file to a
`test_extract` function. However, that is fairly inflexible, and also it turns
out not very much simpler (it doesn't help that `test_extract`
returns `Option<String>`: frankly, it's easier for tests to panic in such
situations).

This commit therefore changes the `test_extract` function signature from:

  ```
  Fn(&str) -> Option<String> + Send + Sync,
  ```

  to:

  ```
  Fn(&Path) -> String + Send + Sync,
  ```

In other words, users now have to both:

  1. Read the contents of a path themselves (but it doesn't necessarily have
     to be the path passed to the function!)
  2. Return a `String` rather than an `Option<String>`.

In practise, most `test_extract` functions can be changed from (roughly):

  ```
  test_extract(|s| { s.lines() ... })
  ```

to:

  ```
  test_extract(|p| { read_to_string(p).lines() })
  ```