Report errors in the test runner as if they were test runners.
Before this commit, if the user made a panic-inducing mistake in, for example,
`test_extract` then they will see text to stderr, but the test counts as
successful.

This commit uses the catch_unwind interface to detect such mistakes and count
them as test failures. This can sometimes be a little counter-intuitive (for
example, panicing inside `test_files` now counts as a test failure, even though
the test hasn't been run), but it does mean that these can't scroll by
harmlessly. Using catch_unwind forces us to make the functions passed to
lang_tester be RefUnwindSafe, which will require a major API change (though,
given the nature of the sorts of functions people are likely to use with
lang_tester, few are likely not to be unwind safe).