Fix stdin close race.
If a test has a "stdin" clause, we close() stdin while still executing the
program to stop it waiting forever for more input that will never come. However,
that means that we closed both the "raw" file handle and dropped the `stdin`
variable which then (sort of silently) closed the file handle a second time.
Since file handles are just integers and can be reused by the operating system,
this led to a highly entertaining race where we closed (say) fd 6, subsequently
fd 6 was reused (perhaps for another child processes's stdout), and then we
dropped/closed fd 6 again -- even though it is in use elsewhere.

The fix for this is to `forget` the `stdin` variable if its been explicitly
dropped.